implémentation Dijkstra non testée

+++ agile/src/main/java/h4131/calculus/InterCompare.java
@@ -4,17 +4,17 @@ import h4131.model.Intersection;
 
 public class InterCompare implements Comparable<InterCompare> {
 
-    public int distance;
+    public double distance;
     public Intersection intersection;
 
-    public InterCompare(int aDistance, Intersection aIntersection) {
+    public InterCompare(double aDistance, Intersection aIntersection) {
         this.distance = aDistance;
         this.intersection = aIntersection;
     }
 
     @Override
     public int compareTo(InterCompare o) {
-        return ((Integer) distance).compareTo(o.distance);
+        return ((Double) distance).compareTo(o.distance);
     }
 
 }
